fix image error
